+++ pushapkscript/test/test_googleplay.py
@@ -1,174 +1,206 @@
 import unittest
 
 import pytest
+from mozapkpublisher.common.apk.checker import AnyPackageNamesCheck, ExpectedPackageNamesCheck
 from mozapkpublisher.push_apk import FileGooglePlayStrings, NoGooglePlayStrings
 from scriptworker.exceptions import TaskVerificationError
-from unittest.mock import MagicMock, patch, ANY
+from unittest.mock import patch, ANY
 
-from pushapkscript.googleplay import publish_to_googleplay, get_service_account, get_certificate_path, \
-    _get_play_config, should_commit_transaction, get_google_play_strings_path, \
+from pushapkscript.exceptions import ConfigValidationError
+from pushapkscript.googleplay import publish_to_googleplay, \
+    should_commit_transaction, get_google_play_strings_path, \
     _check_google_play_string_is_the_only_failed_task, _find_unique_google_play_strings_file_in_dict
 from pushapkscript.test.helpers.mock_file import mock_open, MockFile
 
 
+# TODO: refactor to pytest instead of unittest
 @patch('pushapkscript.googleplay.open', new=mock_open)
 @patch('pushapkscript.googleplay.push_apk')
 class GooglePlayTest(unittest.TestCase):
     def setUp(self):
-        self.context = MagicMock()
-        self.context.config = {
-            'google_play_accounts': {
-                'aurora': {
-                    'service_account': 'aurora_account',
-                    'certificate': '/path/to/aurora.p12',
-                },
-                'beta': {
-                    'service_account': 'beta_account',
-                    'certificate': '/path/to/beta.p12',
-                },
-                'release': {
-                    'service_account': 'release_account',
-                    'certificate': '/path/to/release.p12',
-                },
-                'dep': {
-                    'service_account': 'dummy_dep',
-                    'certificate': '/path/to/dummy_non_p12_file',
-                },
-            },
-            'taskcluster_scope_prefixes': ['project:releng:googleplay:'],
+        self.task_payload = {
+            'google_play_track': 'alpha'
         }
-        self.context.task = {
-            'scopes': ['project:releng:googleplay:release'],
-            'payload': {
-                'google_play_track': 'alpha'
+
+        self.products = {
+            'aurora': {
+                'has_nightly_track': False,
+                'service_account': 'aurora_account',
+                'certificate': '/path/to/aurora.p12',
+                'skip_check_package_names': True,
             },
+            'beta': {
+                'has_nightly_track': False,
+                'service_account': 'beta_account',
+                'certificate': '/path/to/beta.p12',
+                'skip_check_package_names': True,
+            },
+            'release': {
+                'has_nightly_track': False,
+                'service_account': 'release_account',
+                'certificate': '/path/to/release.p12',
+                'skip_check_package_names': True,
+            },
+            'dep': {
+                'has_nightly_track': False,
+                'service_account': 'dummy_dep',
+                'certificate': '/path/to/dummy_non_p12_file',
+                'skip_check_package_names': True,
+                'skip_check_ordered_version_codes': True,
+                'skip_checks_fennec': True,
+            }
         }
         self.apks = [MockFile('/path/to/x86.apk'), MockFile('/path/to/arm_v15.apk')]
 
     def test_publish_config(self, mock_push_apk):
         android_products = ('aurora', 'beta', 'release')
         for android_product in android_products:
-            self.context.task['scopes'] = ['project:releng:googleplay:{}'.format(android_product)]
-            publish_to_googleplay(self.context, self.apks)
+            publish_to_googleplay(self.task_payload, self.products[android_product], self.apks, contact_google_play=True)
 
             mock_push_apk.assert_called_with(
                 apks=[MockFile('/path/to/x86.apk'), MockFile('/path/to/arm_v15.apk')],
                 service_account='{}_account'.format(android_product),
                 google_play_credentials_file=MockFile('/path/to/{}.p12'.format(android_product)),
                 track='alpha',
+                package_names_check=ANY,
                 rollout_percentage=None,
                 google_play_strings=ANY,
                 commit=False,
                 contact_google_play=True,
+                skip_check_multiple_locales=False,
+                skip_check_ordered_version_codes=False,
+                skip_check_same_locales=False,
+                skip_checks_fennec=False,
             )
             _, args = mock_push_apk.call_args
+            package_names_check = args['package_names_check']
             google_play_strings = args['google_play_strings']
+            assert isinstance(package_names_check, AnyPackageNamesCheck)
             assert isinstance(google_play_strings, NoGooglePlayStrings)
 
+    def test_craft_push_config_validates_track(self, mock_push_apk):
+        task_payload_fake_track = {
+            'google_play_track': 'fake'
+        }
+
+        with pytest.raises(TaskVerificationError):
+            publish_to_googleplay(task_payload_fake_track, self.products['release'], self.apks, contact_google_play=True)
+
+        task_payload_nightly_track = {
+            'google_play_track': 'nightly'
+        }
+        with pytest.raises(TaskVerificationError):
+            publish_to_googleplay(task_payload_nightly_track, self.products['release'], self.apks, contact_google_play=True)
+
+        nightly_product_config = {
+            'has_nightly_track': True,
+            'service_account': 'release_account',
+            'certificate': '/path/to/release.p12',
+            'skip_check_package_names': True,
+        }
+        publish_to_googleplay(task_payload_nightly_track, nightly_product_config, self.apks, contact_google_play=True)
+        mock_push_apk.assert_called_once()
+        _, args = mock_push_apk.call_args
+        assert args['track'] == 'nightly'
+
     def test_publish_allows_rollout_percentage(self, mock_push_apk):
-        self.context.task['payload']['google_play_track'] = 'rollout'
-        self.context.task['payload']['rollout_percentage'] = 10
-        publish_to_googleplay(self.context, self.apks)
-        mock_push_apk.assert_called_once_with(
-            apks=ANY,
-            service_account=ANY,
-            google_play_credentials_file=ANY,
-            track='rollout',
-            rollout_percentage=10,
-            google_play_strings=ANY,
-            commit=ANY,
-            contact_google_play=ANY,
-        )
+        task_payload = {
+            'google_play_track': 'rollout',
+            'rollout_percentage': 10
+        }
+        publish_to_googleplay(task_payload, self.products['release'], self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        assert args['track'] == 'rollout'
+        assert args['rollout_percentage'] == 10
 
-    def test_publish_allows_to_contact_google_play_or_not(self, mock_push_apk):
-        publish_to_googleplay(self.context, self.apks)
-        mock_push_apk.assert_called_with(
-            apks=ANY,
-            service_account=ANY,
-            google_play_credentials_file=ANY,
-            track=ANY,
-            rollout_percentage=ANY,
-            google_play_strings=ANY,
-            commit=ANY,
-            contact_google_play=True,
-        )
+    def test_craft_push_config_allows_to_contact_google_play_or_not(self, mock_push_apk):
+        publish_to_googleplay(self.task_payload, self.products['aurora'], self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        assert args['contact_google_play'] is True
 
-        self.context.config['do_not_contact_google_play'] = True
-        publish_to_googleplay(self.context, self.apks)
-        mock_push_apk.assert_called_with(
-            apks=ANY,
-            service_account=ANY,
-            google_play_credentials_file=ANY,
-            track=ANY,
-            rollout_percentage=ANY,
-            google_play_strings=ANY,
-            commit=ANY,
-            contact_google_play=False,
-        )
+        publish_to_googleplay(self.task_payload, self.products['aurora'], self.apks, False)
+        _, args = mock_push_apk.call_args
+        assert args['contact_google_play'] is False
 
-    def test_publish_allows_committing_apks(self, mock_push_apk):
-        self.context.task['payload']['commit'] = True
-        publish_to_googleplay(self.context, self.apks)
+    def test_craft_push_config_skip_checking_multiple_locales(self, mock_push_apk):
+        product_config = {
+            'has_nightly_track': False,
+            'service_account': 'product',
+            'certificate': '/path/to/product.p12',
+            'skip_check_package_names': True,
+            'skip_check_multiple_locales': True,
+        }
+        publish_to_googleplay(self.task_payload, product_config, self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        assert args['skip_check_multiple_locales'] is True
 
-        mock_push_apk.assert_called_with(
-            apks=ANY,
-            service_account=ANY,
-            google_play_credentials_file=ANY,
-            track=ANY,
-            rollout_percentage=ANY,
-            google_play_strings=ANY,
-            commit=True,
-            contact_google_play=ANY,
-        )
+    def test_craft_push_config_skip_checking_same_locales(self, mock_push_apk):
+        product_config = {
+            'has_nightly_track': False,
+            'service_account': 'product',
+            'certificate': '/path/to/product.p12',
+            'skip_check_package_names': True,
+            'skip_check_same_locales': True,
+        }
+        publish_to_googleplay(self.task_payload, product_config, self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        assert args['skip_check_same_locales'] is True
 
-    def test_publish_raises_error_when_android_product_is_not_part_of_config(self, _):
-        self.context.task['scopes'] = ['project:releng:googleplay:non_existing_android_product']
-        self.assertRaises(TaskVerificationError, publish_to_googleplay, self.context, self.apks)
+    def test_craft_push_config_expect_package_names(self, mock_push_apk):
+        product_config = {
+            'has_nightly_track': False,
+            'service_account': 'product',
+            'certificate': '/path/to/product.p12',
+            'expected_package_names': ['org.mozilla.focus', 'org.mozilla.klar']
+        }
+        publish_to_googleplay(self.task_payload, product_config, self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        package_names_check = args['package_names_check']
+        assert isinstance(package_names_check, ExpectedPackageNamesCheck)
+        assert package_names_check.expected_product_types == ['org.mozilla.focus', 'org.mozilla.klar']
 
-    def test_publish_raises_error_when_google_play_accounts_does_not_exist(self, _):
-        del self.context.config['google_play_accounts']
-        self.assertRaises(TaskVerificationError, publish_to_googleplay, self.context, self.apks)
+    def test_craft_push_config_allows_committing_apks(self, mock_push_apk):
+        task_payload = {
+            'google_play_track': 'production',
+            'commit': True
+        }
+        publish_to_googleplay(task_payload, self.products['release'], self.apks, contact_google_play=True)
+        _, args = mock_push_apk.call_args
+        assert args['commit'] is True
 
     def test_publish_updates_google_strings_from_file(self, mock_push_apk):
-        publish_to_googleplay(self.context, self.apks,
+        publish_to_googleplay(self.task_payload, self.products['release'], self.apks, contact_google_play=True,
                               google_play_strings_file=MockFile('/path/to/google_play_strings.json'))
         _, args = mock_push_apk.call_args
         google_play_strings = args['google_play_strings']
         assert isinstance(google_play_strings, FileGooglePlayStrings)
         assert google_play_strings.file.name == '/path/to/google_play_strings.json'
 
-    def test_get_service_account(self, _):
-        self.assertEqual(get_service_account(self.context, 'aurora'), 'aurora_account')
-        self.assertEqual(get_service_account(self.context, 'beta'), 'beta_account')
-        self.assertEqual(get_service_account(self.context, 'release'), 'release_account')
 
-    def test_get_certificate_path(self, _):
-        self.assertEqual(get_certificate_path(self.context, 'aurora'), '/path/to/aurora.p12')
-        self.assertEqual(get_certificate_path(self.context, 'beta'), '/path/to/beta.p12')
-        self.assertEqual(get_certificate_path(self.context, 'release'), '/path/to/release.p12')
+def test_package_name_validation():
+    task_payload = {
+        'google_play_track': 'production'
+    }
+    product_config = {
+        'has_nightly_track': False
+    }
+    with pytest.raises(ConfigValidationError):
+        publish_to_googleplay(task_payload, product_config, [], contact_google_play=True)
 
-    def test_get_play_config(self, _):
-        self.assertEqual(_get_play_config(self.context, 'aurora'), {
-            'service_account': 'aurora_account', 'certificate': '/path/to/aurora.p12'
-        })
 
-        self.assertRaises(TaskVerificationError, _get_play_config, self.context, 'non-existing-android-product')
+def test_should_commit_transaction():
+    task_payload = {
+        'commit': True
+    }
+    assert should_commit_transaction(task_payload) is True
 
-        class FakeContext:
-            config = {}
+    task_payload = {
+        'commit': False
+    }
+    assert should_commit_transaction(task_payload) is False
 
-        context_without_any_account = FakeContext()
-        self.assertRaises(TaskVerificationError, _get_play_config, context_without_any_account, 'whatever-android-product')
-
-    def test_should_commit_transaction(self, _):
-        self.context.task['payload']['commit'] = True
-        self.assertTrue(should_commit_transaction(self.context))
-
-        self.context.task['payload']['commit'] = False
-        self.assertFalse(should_commit_transaction(self.context))
-
-        del self.context.task['payload']['commit']
-        self.assertFalse(should_commit_transaction(self.context))
+    task_payload = {}
+    assert should_commit_transaction(task_payload) is False
 
 
 def test_get_google_play_strings_path():